ui/ui_list.c, ui_account.c: add support for scrolling (WIP)
It works, but if we overshoot, the distance overshot is recorded, and, if
not lifting the finger off the screen, one has to cover the distance
overshot when trying to scroll back. While this isn't "wrong", it may
feel unusual.

E.g., the following Android app does it like this:
- Heise News

While these respond immediately if the scroll direction is reversed:
- Chrome browser
- INFOBAE
- Avherald (article display)
- HSBC Argentina